+++ src/Models/FloatingIps/FloatingIp.php
@@ -177,7 +177,7 @@ class FloatingIp extends Model
         if (!HetznerAPIClient::hasError($response)) {
             return APIResponse::create([
                 'action' => Action::parse(json_decode((string)$response->getBody())->action)
-            ]);
+            ], $response->getHeaders());
         }
     }
 
@@ -199,7 +199,7 @@ class FloatingIp extends Model
         if (!HetznerAPIClient::hasError($response)) {
             return APIResponse::create([
                 'action' => Action::parse(json_decode((string)$response->getBody())->action)
-            ]);
+            ], $response->getHeaders());
         }
     }
 
@@ -216,7 +216,7 @@ class FloatingIp extends Model
         if (!HetznerAPIClient::hasError($response)) {
             return APIResponse::create([
                 'action' => Action::parse(json_decode((string)$response->getBody())->action)
-            ]);
+            ], $response->getHeaders());
         }
     }
 
@@ -240,7 +240,7 @@ class FloatingIp extends Model
         if (!HetznerAPIClient::hasError($response)) {
             return APIResponse::create([
                 'action' => Action::parse(json_decode((string)$response->getBody())->action)
-            ]);
+            ], $response->getHeaders());
         }
     }