+++ src/Models/SSHKeys/SSHKeys.php
@@ -55,7 +55,7 @@ class SSHKeys extends Model
      * @return array
      * @throws \LKDev\HetznerCloud\APIException
      */
-    public function all(RequestOpts $requestOpts): array
+    public function all(RequestOpts $requestOpts = null): array
     {
         if ($requestOpts == null) {
             $requestOpts = new RequestOpts();